requirements.txt
aiofiles==24.1.0
aiohttp==3.12.15
beautifulsoup4==4.13.5
cachetools==6.2.0
chromadb==1.0.20
cmudict==1.1.1
cryptography==45.0.7
discord.py==2.6.3
python-docx==1.1.2
ebooklib==0.19
fake_useragent==2.2.0
faster_whisper==1.2.0
pymupdf==1.26.3
Flask==3.1.2
g2p_en==2.1.0
httpx[http2]==0.28.1
brotli==1.1.0
kokoro_onnx==0.4.9
misaki==0.9.4
nltk==3.9.1
numpy==2.3.2
onnxruntime==1.22.1
openai==1.107.0
pdf2image==1.17.0
phonemizer==3.3.0
Pillow==11.3.0
playwright==1.55.0
prometheus-client==0.22.1
psutil==7.0.0
pydub==0.25.1
PyPDF2==3.0.1
pytesseract==0.3.13
pytest==8.4.1
pytest-asyncio==0.23.8
python-dotenv==1.1.1
python-magic==0.4.27
reportlab==4.4.3
requests==2.32.5
rich==14.1.0
scipy==1.16.1
sentence-transformers==5.1.0
soundfile==0.13.1
torch==2.3.1
tqdm==4.67.1
trafilatura==2.0.0
urllib3==2.5.0
